+++ src/stable/stable.ts
@@ -1,0 +1,3 @@
+import { Liquidity } from "../liquidity";
+
+export class Stable extends Liquidity {}